SQL Extraction: minor VS Code refactoring (#140)
